requirements.txt
# Core dependencies
boto3>=1.28.0
psycopg2-binary>=2.9.5
sqlparse>=0.4.3
markdown>=3.4.1
jinja2>=3.1.2

# AWS Bedrock related
botocore>=1.31.0
anthropic>=0.5.0

# API and serverless
requests>=2.28.2
fastapi>=0.95.0
pydantic>=1.10.7
uvicorn[standard]>=0.20.0
serverless-wsgi>=3.0.1

# Advanced RAG techniques
numpy>=1.24.2
python-Levenshtein>=0.20.9

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
colorama>=0.4.6

# MySQL support
mysql-connector-python>=8.0.33

# Query analysis dependencies
sql_metadata>=2.5.0
sqlglot>=23.6.0
pandas>=2.0.0